requirements.txt
Django==2.2
PyYAML
mysqlclient
wikipedia-API
facebook-sdk
django-import-export

#phaser

enum-compat==0.0.2
eventlet==0.21.0
python-engineio
python-socketio
pytz==2018.7
six==1.10.0
